+++ crates/module-system/module-implementations/sov-chain-state/tests/integration-tests/dynamic_gas_update.rs
@@ -106,10 +106,12 @@ fn test_gas_price_increases_if_gas_used_exceeds_gas_target() {
         }),
     });
 
-    let result =
-        runner.execute(user.create_plain_message::<RT, ValueSetter<S>>(
-            sov_value_setter::CallMessage::SetValue(10),
-        ));
+    let result = runner.execute(user.create_plain_message::<RT, ValueSetter<S>>(
+        sov_value_setter::CallMessage::SetValue {
+            value: 10,
+            gas: None,
+        },
+    ));
 
     assert_eq!(result.batch_receipts.len(), 1);
     let gas_price = result.batch_receipts[0].inner.gas_price.clone();
@@ -154,10 +156,12 @@ fn test_gas_price_decreases_if_gas_used_is_below_gas_target() {
         }),
     });
 
-    let result =
-        runner.execute(user.create_plain_message::<RT, ValueSetter<S>>(
-            sov_value_setter::CallMessage::SetValue(10),
-        ));
+    let result = runner.execute(user.create_plain_message::<RT, ValueSetter<S>>(
+        sov_value_setter::CallMessage::SetValue {
+            value: 10,
+            gas: None,
+        },
+    ));
 
     assert_eq!(result.batch_receipts.len(), 1);
     let gas_price = result.batch_receipts[0].inner.gas_price.clone();